+++ Snippets/FileBasedRouting/FIleBasedRouting.sln
@@ -1,20 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FIleBasedRouting_1", "FIleBasedRouting_1\FIleBasedRouting_1.csproj", "{FE2253DC-45F8-4254-B6AA-11AFE9EACEB5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{FE2253DC-45F8-4254-B6AA-11AFE9EACEB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FE2253DC-45F8-4254-B6AA-11AFE9EACEB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FE2253DC-45F8-4254-B6AA-11AFE9EACEB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FE2253DC-45F8-4254-B6AA-11AFE9EACEB5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
