--- Snippets/FileBasedRouting/FIleBasedRouting_1/App.config
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6"/>
-    </startup>
-</configuration>
