+++ Snippets/FileBasedRouting/FIleBasedRouting_1/endpoints-by-assembly.xml
@@ -5,7 +5,8 @@
       <!-- Define that endpointA can handle all commands in assembly "MyApp.Contracts" -->
       <commands assembly="MyApp.Contracts" />
       <!-- Define that endpointA can handle all events in assembly "MyApp.Contracts" within the namespace "Events" -->
-      <events assembly="MyApp.Contracts" namespace="Events" />
+      <events assembly="MyApp.Contracts"
+              namespace="Events" />
     </handles>
   </endpoint>
 </endpoints>
