+++ nservicebus/messaging/file-based-routing.md
@@ -13,7 +13,8 @@ related:
 - nservicebus/messaging/routing
 ---
 
-NOTE: Centralized file-based routing is an experimental project hosted on ParticularLabs. Particular Software does not provide any support guarantee for such projects.
+NOTE: Centralized file-based routing is an experimental project hosted on [ParticularLabs](https://github.com/ParticularLabs). Particular Software does not provide any support guarantee for such projects.
+
 
 ## Concept 
 
@@ -26,7 +27,7 @@ Enable the feature in via the routing configuration:
 
 snippet:Enable
 
-This will configure the endpoint to look for a `endpoints.xml` file in the endpoints [base directory](https://msdn.microsoft.com/en-us/library/system.appdomain.basedirectory(v=vs.110).aspx).
+This will configure the endpoint to look for a `endpoints.xml` file in the endpoints [base directory](https://msdn.microsoft.com/en-us/library/system.appdomain.basedirectory.aspx).
 
 It is possible to configure a different relative or absolute file path, e.g.:
 
@@ -38,12 +39,13 @@ Create a new XML file named `endpoints.xml` and include it on every endpoint usi
 
 snippet:EndpointsByType
 
-The `type` attribute needs to provide the [Assembly Qualified Type Name](https://msdn.microsoft.com/en-us/library/system.type.assemblyqualifiedname(v=vs.110).aspx).
+The `type` attribute needs to provide the [Assembly Qualified Type Name](https://msdn.microsoft.com/en-us/library/system.type.assemblyqualifiedname.aspx).
 Make sure that the routing file is copied to the binaries output folder.
 
 Instead of defining every single message type, it's also possible to configure entire assemblies or namespaces in bulk:
 
 snippet:EndpointsByAssembly
+
 
 ### Updating the routing configuration
 
@@ -53,9 +55,10 @@ The routing configuration is read every 30 seconds, therefore the topology can c
 ### Sharing the routing file
 
 In order to allow centralized configuration of the routing file, the file needs to be shared with the endpoints. This can be done in various ways, e.g.
-* Make the file available via a shared network folder.
-* Distribute the file as a part of deployment process.
-* Include the file in the project/solution and its build artifacts. Note: This approach does not allow for a centralized routing file management out of the box.
+
+ * Make the file available via a shared network folder.
+ * Distribute the file as a part of deployment process.
+ * Include the file in the project/solution and its build artifacts. Note: This approach does not allow for a centralized routing file management out of the box.
 
 
 ## Scaling out
